Corrigi erros no BattleShipTest, gerei report de testes e confirmei a checklist

+++ src/test/java/iscteiul/ista/battleship/ShipTest.java
@@ -42,7 +42,7 @@ class ShipTest {
 
         @Override
         public Integer getSize() {
-            return size;
+            return (Integer) size;
         }
     }
